requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.0.2
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.5
pygame==2.1.2
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.10
Werkzeug==2.0.2
WTForms==3.0.1
